Move around scripting utils
