+++ src/sql/workbench/browser/scriptingUtils.ts
